# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Support Ticketing System API
  version: 1.0.0
  description: |
    Comprehensive API for a support ticketing system with users, tickets, comments, attachments, knowledge base, notifications, and real-time messaging with threaded conversations.

    ## Key Features

    - **Multi-channel Messaging**: Support for direct messages, group conversations, and ticket-based discussions
    - **Threaded Replies**: Full support for nested message replies in both conversations and ticket comments
    - **Real-time Notifications**: Dual-channel notifications via webhooks and WebSocket/Socket.IO
    - **Message Read Status**: WhatsApp-like read receipts with per-user message tracking
    - **Comprehensive Ticketing**: Full lifecycle management of support tickets with status tracking
    - **Knowledge Base**: Article management with tagging and search capabilities
    - **File Attachments**: Support for attaching files to tickets and messages
    - **User Management**: Role-based access control and user administration

    ## Real-time Notifications & Webhooks

    The API supports comprehensive real-time notifications through **dual channels**:

    ### Webhooks (HTTP-based)
    - External integrations and reliable delivery
    - Configurable per user with retry logic
    - Includes full entity data for UI updates

    ### WebSockets/Socket.IO (Real-time)
    - Instant bidirectional communication
    - Room-based messaging for targeted updates
...

...

...

...

...

...

...
    {
...

...

...
servers:
...
paths:
...
  /api/notifications/webhooks/test:
    post:
      summary: Test webhook configuration
      tags: [webhooks]
      security:
        - bearerAuth: []
      description: Send a test webhook to the current user's configured webhook URL
      responses:
        '200':
          description: Test webhook sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: test webhook sent
        '400':
          description: No webhook URL configured
        '500':
          description: Failed to send test webhook
...
  /api/testing/:
    get:
      tags: [testing]
      summary: List testing sessions
      parameters:
        - name: ticket_id
          in: query
          schema:
            type: string
          description: Filter by ticket ID
        - name: user_id
          in: query
          schema:
            type: string
          description: Filter by user ID
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, passed, failed]
          description: Filter by status
      responses:
        '200':
          description: List of testing sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Testing'
    post:
      tags: [testing]
      summary: Create a new testing session
      description: Can only be created for tickets with status 'resolved'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewTesting'
      responses:
        '201':
          description: Testing session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testing'
        '400':
          description: Invalid request or ticket not eligible for testing
        '401':
          description: Unauthorized
  /api/testing/{id_}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [testing]
      summary: Get testing session by ID
      responses:
        '200':
          description: Testing session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testing'
        '404':
          description: Testing session not found
    put:
      tags: [testing]
      summary: Update testing session
      description: Only the tester can update their testing session. When status is set to 'passed' or 'failed', the associated ticket status is updated and the assignee is notified.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestingUpdate'
      responses:
        '200':
          description: Testing session updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Testing'
        '403':
          description: Forbidden - only tester can update
        '404':
          description: Testing session not found
    delete:
      tags: [testing]
      summary: Delete testing session
      description: Only the tester can delete their testing session
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Testing session deleted
        '403':
          description: Forbidden - only tester can delete
        '404':
          description: Testing session not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id_
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Testing:
      type: object
      properties:
        id: { type: string }
        ticket_id: { type: string }
        user_id: { type: string }
        status: { type: string, enum: [pending, in_progress, passed, failed] }
        test_type: { type: string, enum: [manual, unit, integration] }
        results: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NewTesting:
      type: object
      required: [ticket_id]
      properties:
        ticket_id: { type: string }
        status: { type: string, enum: [pending, in_progress, passed, failed] }
        test_type: { type: string, enum: [manual, unit, integration] }
        results: { type: string }
    TestingUpdate:
      type: object
      properties:
        status: { type: string, enum: [pending, in_progress, passed, failed] }
        test_type: { type: string, enum: [manual, unit, integration] }
        results: { type: string }
...